Extract safeCoreSDK deployment override logic into utils
